Adds Comparable module. Passes tests
